+++ src/TorchSharp/TorchAudio/IAudioBackend.cs
@@ -9,12 +9,15 @@ namespace TorchSharp
     {
         public static partial class backend
         {
+            /// <summary>
+            /// Audio I/O backend
+            /// </summary>
             public interface IAudioBackend
             {
                 (torch.Tensor, int) load(
                     string filepath,
-                    int frame_offset = 0,
-                    int num_frames = -1,
+                    long frame_offset = 0,
+                    long num_frames = -1,
                     bool normalize = true,
                     bool channels_first = true,
                     AudioFormat? format = null);
